backend/requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0